refactor: split web lint warning modules
