+++ packages/web/components/detail/IssueActionSheet.tsx
@@ -11,7 +11,8 @@ import { IssueActionsMenu } from "./IssueActionsMenu";
 import { IssueDesktopActions } from "./IssueDesktopActions";
 import { IssueReassignSheet } from "./IssueReassignSheet";
 import { AutoLaunchTrigger } from "./AutoLaunchTrigger";
-import { closeIssue, reassignIssueAction } from "@/lib/actions/issues";
+import { closeIssue } from "@/lib/actions/issues";
+import { reassignIssueAction } from "@/lib/actions/issues-reassign";
 import { endSession } from "@/lib/actions/launch";
 import { getComments } from "@/lib/actions/comments";
 import { listReposAction } from "@/lib/actions/drafts";
