+++ packages/web/lib/actions/issues.ts
@@ -8,7 +8,6 @@ import {
   updateIssue as coreUpdateIssue,
   closeIssue as coreCloseIssue,
   reopenIssue as coreReopenIssue,
-  reassignIssue as coreReassignIssue,
   addLabel as coreAddLabel,
   removeLabel as coreRemoveLabel,
   addComment as coreAddComment,
@@ -17,7 +16,6 @@ import {
   withIdempotency,
   DuplicateInFlightError,
   formatErrorForUser,
-  type ReassignResult,
 } from "@issuectl/core";
 import { revalidateSafely } from "@/lib/revalidate";
 import { notifyNewIssue } from "@/lib/push/notifications";
@@ -266,86 +264,4 @@ export async function toggleLabel(data: {
   }
   const { stale } = revalidateSafely(`/issues/${owner}/${repo}/${number}`, "/");
   return { success: true, ...(stale ? { cacheStale: true as const } : {}) };
-}
-
-export async function reassignIssueAction(
-  oldRepoId: number,
-  issueNumber: number,
-  newRepoId: number,
-  idempotencyKey?: string,
-): Promise<
-  | {
-      success: true;
-      newIssueNumber: number;
-      newOwner: string;
-      newRepo: string;
-      cleanupWarning?: string;
-      cacheStale?: true;
-    }
-  | { success: false; error: string }
-> {
-  if (
-    typeof oldRepoId !== "number" ||
-    !Number.isInteger(oldRepoId) ||
-    oldRepoId <= 0
-  ) {
-    return { success: false, error: "oldRepoId must be a positive integer" };
-  }
-  if (
-    typeof newRepoId !== "number" ||
-    !Number.isInteger(newRepoId) ||
-    newRepoId <= 0
-  ) {
-    return { success: false, error: "newRepoId must be a positive integer" };
-  }
-  if (
-    typeof issueNumber !== "number" ||
-    !Number.isFinite(issueNumber) ||
-    issueNumber <= 0
-  ) {
-    return {
-      success: false,
-      error: "issueNumber must be a positive integer",
-    };
-  }
-  if (oldRepoId === newRepoId) {
-    return { success: false, error: "Cannot re-assign to the same repo" };
-  }
-
-  let result: ReassignResult;
-  try {
-    const db = getDb();
-
-    const runReassign = async () => {
-      return withAuthRetry((octokit) =>
-        coreReassignIssue(db, octokit, oldRepoId, issueNumber, newRepoId),
-      );
-    };
-
-    // Idempotency guard: a retry after partial failure (new issue
-    // created but old not yet closed) replays the stored result
-    // rather than creating a duplicate issue on the target repo.
-    result = idempotencyKey
-      ? await withIdempotency(db, "reassign-issue", idempotencyKey, runReassign)
-      : await runReassign();
-  } catch (err) {
-    if (err instanceof DuplicateInFlightError) {
-      return {
-        success: false,
-        error: "This issue is already being re-assigned — please wait.",
-      };
-    }
-    console.error("[issuectl] Failed to re-assign issue:", err);
-    return { success: false, error: formatErrorForUser(err) };
-  }
-
-  const { stale } = revalidateSafely("/");
-  return {
-    success: true,
-    newIssueNumber: result.newIssueNumber,
-    newOwner: result.newOwner,
-    newRepo: result.newRepo,
-    ...(result.cleanupWarning ? { cleanupWarning: result.cleanupWarning } : {}),
-    ...(stale ? { cacheStale: true as const } : {}),
-  };
 }
